All tileset func
Added tileset and built a little testing area. Removed test enemies from my scene. File is a zip folder of the asset folder as there are to many files to add. May try to uncompress later however it will just be easier if you uncompress and swap for your asset folder.